refactor: move SimpleMemberCard to its own folder, create a reusable skeleton
