+++ src/components/frame/v5/pages/MembersPage/partials/MembersTabContent/MembersTabContent.tsx
@@ -8,12 +8,12 @@ import { useColonyContext } from '~context/ColonyContext/ColonyContext.ts';
 import { useFilterContext } from '~context/FilterContext/FilterContext.ts';
 import { useSearchContext } from '~context/SearchContext/SearchContext.ts';
 import useCopyToClipboard from '~hooks/useCopyToClipboard.ts';
-import { SpinnerLoader } from '~shared/Preloaders/index.ts';
 import { formatText } from '~utils/intl.ts';
 import EmptyContent from '~v5/common/EmptyContent/index.ts';
 import MemberCard from '~v5/common/MemberCard/index.ts';
-import SimpleMemberCard from '~v5/common/MemberCard/SimpleMemberCard.tsx';
 import MemberCardList from '~v5/common/MemberCardList/index.ts';
+import SimpleMemberCard from '~v5/common/SimpleMemberCard/SimpleMemberCard.tsx';
+import SimpleMemberCardSkeleton from '~v5/common/SimpleMemberCard/SimpleMemberCardSkeleton.tsx';
 import { TextButton } from '~v5/shared/Button/index.ts';
 
 import { useMembersTabContentItems } from './hooks.tsx';
@@ -106,13 +106,11 @@ const MembersTabContent: FC<PropsWithChildren<MembersTabContentProps>> = ({
             <EmptyContent {...emptyContentProps} withBorder />
           )}
         {isLoading && (
-          <div
-            className={clsx('flex w-full items-center justify-center', {
-              'mt-6': !!items.length,
-            })}
-          >
-            <SpinnerLoader />
-          </div>
+          <MemberCardList isSimple={withSimpleCards}>
+            {[...Array(16).keys()].map((key) => (
+              <SimpleMemberCardSkeleton key={key} />
+            ))}
+          </MemberCardList>
         )}
         {loadMoreButtonProps && (
           <div className="mt-6 flex w-full justify-center">
